refactor(markdown): use list names instead of IDs for card creation
Replace ListID with ListName to handle cases where lists are created in the same operation. Sentinel IDs for new lists cannot be resolved during card creation so we get the lists and loop over them until we find the name and then use that id when selecting which list to put the card in.